requirements.txt
customtkinter
watchdog
pystray
pillow
requests
